@@ -51,7 +51,7 @@ class HandleInertiaRequests extends Middleware
                     ['label' => '100', 'value' => 100],
                 ],
             ],
-            'locale' => function () {                
+            'locale' => function () {
                 if(session()->has('locale')) {
                     app()->setLocale(session('locale'));
                 }
