@@ -2,6 +2,8 @@
 
 namespace App\Http\Requests\Permission;
 
+
+use Illuminate\Validation\Rule;
 use Illuminate\Foundation\Http\FormRequest;
 
 class PermissionUpdateRequest extends FormRequest
@@ -23,8 +25,16 @@ class PermissionUpdateRequest extends FormRequest
      */
     public function rules()
     {
-        return [
-            'name' => 'required|unique:permissions,name,' . $this->permission->id,
-        ];
+       // return [
+           // 'name' => 'required|unique:permissions,name,' . $this->permission->id,
+       // ];
+       return [
+        'name' => [
+            'required',
+            Rule::unique('permissions', 'name')->ignore(
+                $this->route('permission')?->id // الحل الأكثر أمانًا
+            ),
+        ],
+    ];
     }
 }
