@@ -2,8 +2,8 @@
 
 namespace App\Models;
 
-use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
+use Illuminate\Database\Eloquent\Factories\HasFactory;
 
 class Answer extends Model
 {
@@ -16,13 +16,46 @@ class Answer extends Model
         'is_correct',
     ];
 
+    /**
+     * Get the question that owns the answer.
+     *
+     * @return \Illuminate\Database\Eloquent\Relations\BelongsTo
+     */
     public function question()
     {
         return $this->belongsTo(Question::class);
     }
 
+    /**
+     * Get all translations for the answer.
+     *
+     * @return \Illuminate\Database\Eloquent\Relations\HasMany
+     */
     public function translations()
     {
-        return $this->hasMany(AnswerTranslation::class);
+        return $this->hasMany(\App\Models\AnswerTranslation::class);
+    }
+
+    /**
+     * Get the answer in a specific locale.
+     *
+     * @param string|null $locale
+     * @return string|null
+     */
+    public function getTranslation(?string $locale = null): ?string
+    {
+        $locale = $locale ?? app()->getLocale();
+
+        if ($locale === 'en') {
+            return $this->answer_en;
+        }
+
+        if ($locale === 'ar') {
+            return $this->answer_ar;
+        }
+
+        return $this->translations()
+            ->where('locale', $locale)
+            ->value('text');
     }
 }
