@@ -4,6 +4,8 @@ namespace App\Models;
 
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
+use Illuminate\Database\Eloquent\Relations\BelongsTo;
+use Illuminate\Database\Eloquent\Relations\HasMany;
 
 class Section extends Model
 {
@@ -11,12 +13,13 @@ class Section extends Model
 
     protected $guarded = [];
 
-    public function course(){
-    return $this->belongsTo(Course::class);
+    public function course(): BelongsTo
+    {
+        return $this->belongsTo(Course::class);
+    }
 
-}
-public function lectures()
-{
-    return $this->hasMany(Lecture::class);
-}
+    public function lectures(): HasMany
+    {
+        return $this->hasMany(Lecture::class);
+    }
 }
