@@ -41,7 +41,13 @@ return [
     |
     */
 
-    'debug' => (bool) env('APP_DEBUG', false),
+    'debug' => env('APP_DEBUG', false),
+    /*
+    |--------------------------------------------------------------------------
+    | Error Reporting
+    |--------------------------------------------------------------------------
+    */
+    'error_reporting' => E_ALL & ~E_DEPRECATED,
 
     /*
     |--------------------------------------------------------------------------
