@@ -2,6 +2,8 @@
 
 namespace Database\Factories;
 
+use App\Models\Course;
+use App\Models\Section;
 use Illuminate\Database\Eloquent\Factories\Factory;
 
 /**
@@ -16,8 +18,28 @@ class QuizFactory extends Factory
      */
     public function definition(): array
     {
+        $course = Course::inRandomOrder()->first() ?? Course::factory()->create();
+        $section = Section::where('course_id', $course->id)->inRandomOrder()->first()
+                  ?? Section::factory()->create(['course_id' => $course->id]);
+
         return [
-            //
+            'title' => $this->faker->sentence,
+            'description' => $this->faker->paragraph,
+            'time_limit' => $this->faker->numberBetween(10, 60),
+            'passing_score' => $this->faker->numberBetween(60, 80),
+            'is_active' => $this->faker->boolean(80),
+            'course_id' => $course->id,
+            'section_id' => $section->id,
         ];
+    }
+
+    /**
+     * تعيين الاختبار كنشط
+     */
+    public function active(): static
+    {
+        return $this->state(fn (array $attributes) => [
+            'is_active' => true,
+        ]);
     }
 }
