@@ -6,7 +6,7 @@ import InputLabel from "@/Components/InputLabel.vue";
 import PrimaryButton from "@/Components/PrimaryButton.vue";
 import DangerButton from "@/Components/DangerButton.vue";
 import TextInput from "@/Components/TextInput.vue";
-import { Head, Link, useForm } from "@inertiajs/vue3";
+import { Head, Link, useForm, usePage } from "@inertiajs/vue3";
 import AuntheticationIllustration from "@/Components/AuntheticationIllustration.vue";
 
 defineProps({
@@ -16,7 +16,7 @@ defineProps({
 
 const form = useForm({
     email: "",
-    password: "",
+    password: "",  // Make sure password is initialized
     remember: false,
 });
 
@@ -24,31 +24,49 @@ const submit = () => {
     form.post(route("login"), {
         onFinish: () => form.reset("password"),
     });
-
 };
+
 const loginByGoogle = () => {
-    //form.get(route("google.login"), {});
-    window.location.href = route('google.login');
-
+    try {
+        if (!route('google.login')) {
+            throw new Error('Google login route not defined');
+        }
+        window.location.href = route('google.login');
+    } catch (error) {
+        console.error('Google login error:', error);
+        form.errors.email = error.message;
+    }
 };
 
+// استبدال دالة lang() بالترجمات من Laravel
+const __ = (key) => {
+  const translations = {
+    'Login': 'تسجيل الدخول',
+    'Email': 'البريد الإلكتروني',
+    'Password': 'كلمة المرور',
+    'Remember me': 'تذكرني',
+    'Forgot your password?': 'نسيت كلمة المرور؟',
+    'Log in': 'تسجيل الدخول'
+  };
+  return translations[key] || key;
+};
 </script>
 
 <template>
     <GuestLayout>
-        <Head :title="lang().label.login" />
+        <Head :title="__('Login')" />
         <template #illustration>
             <AuntheticationIllustration type="login" class="w-72 h-auto" />
         </template>
         <div v-if="status" class="mb-4 font-medium text-sm text-green-600">
             {{ status }}
 
-           
+
         </div>
 
         <form @submit.prevent="submit">
             <div>
-                <InputLabel for="email" :value="lang().label.email" />
+                <InputLabel for="email" :value="__('Email')" />
                 <TextInput
                     id="email"
                     type="email"
@@ -57,14 +75,14 @@ const loginByGoogle = () => {
                     required
                     autofocus
                     autocomplete="username"
-                    :placeholder="lang().placeholder.email"
+                    :placeholder="__('Email')"
                     :error="form.errors.email"
                 />
                 <InputError class="mt-2" :message="form.errors.email" />
             </div>
 
             <div class="mt-4">
-                <InputLabel for="password" :value="lang().label.password" />
+                <InputLabel for="password" :value="__('Password')" />
                 <TextInput
                     id="password"
                     type="password"
@@ -72,7 +90,7 @@ const loginByGoogle = () => {
                     v-model="form.password"
                     required
                     autocomplete="current-password"
-                    :placeholder="lang().placeholder.password"
+                    :placeholder="__('Password')"
                     :error="form.errors.password"
                 />
                 <InputError class="mt-2" :message="form.errors.password" />
@@ -83,7 +101,7 @@ const loginByGoogle = () => {
                     <Checkbox name="remember" v-model:checked="form.remember" />
                     <span
                         class="ml-2 text-sm text-slate-600 dark:text-slate-400"
-                        >{{ lang().label.remember_me }}</span
+                        >{{ __('Remember me') }}</span
                     >
                 </label>
             </div>
@@ -93,7 +111,7 @@ const loginByGoogle = () => {
                     :href="route('password.request')"
                     class="underline text-sm text-slate-600 dark:text-slate-400 hover:text-slate-900 dark:hover:text-slate-100 rounded-md focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-primary dark:focus:ring-offset-slate-800"
                 >
-                    {{ lang().label.lost_password }}
+                    {{ __('Forgot your password?') }}
                 </Link>
 
                 <PrimaryButton
@@ -103,8 +121,8 @@ const loginByGoogle = () => {
                 >
                     {{
                         form.processing
-                            ? lang().button.login + "..."
-                            : lang().button.login
+                            ? __('Log in') + "..."
+                            : __('Log in')
                     }}
                 </PrimaryButton>
                 <form @submit.prevent="loginByGoogle">
