@@ -10,10 +10,18 @@ import {
 import { Head, Link } from "@inertiajs/vue3";
 
 const props = defineProps({
-    users: Number,
-    roles: Number,
-    permissions: Number,
-    courses: Number,
+    users: {
+        type: Number,
+        default: 0
+    },
+    roles: {
+        type: Number,
+        default: 0
+    },
+    permissions: {
+        type: Number,
+        default: 0
+    }
 });
 </script>
 
@@ -32,7 +40,7 @@ const props = defineProps({
                         <div class="flex flex-col">
                             <p class="text-4xl font-bold">{{ props.users }}</p>
                             <p class="text-md md:text-lg uppercase">
-                                {{ lang().label.user }}
+                                Users
                             </p>
                         </div>
                         <div>
@@ -46,7 +54,7 @@ const props = defineProps({
                             :href="route('user.index')"
                             class="flex justify-between items-center"
                         >
-                            <p>{{ lang().label.more }}</p>
+                            <p>More</p>
                             <ChevronRightIcon class="w-5 h-5" />
                         </Link>
                     </div>
@@ -58,7 +66,7 @@ const props = defineProps({
                         <div class="flex flex-col">
                             <p class="text-4xl font-bold">{{ props.roles }}</p>
                             <p class="text-md md:text-lg uppercase">
-                                {{ lang().label.role }}
+                                Roles
                             </p>
                         </div>
                         <div>
@@ -72,7 +80,7 @@ const props = defineProps({
                             :href="route('role.index')"
                             class="flex justify-between items-center"
                         >
-                            <p>{{ lang().label.more }}</p>
+                            <p>More</p>
                             <ChevronRightIcon class="w-5 h-5" />
                         </Link>
                     </div>
@@ -86,7 +94,7 @@ const props = defineProps({
                                 {{ props.permissions }}
                             </p>
                             <p class="text-md md:text-lg uppercase">
-                                {{ lang().label.permission }}
+                                Permissions
                             </p>
                         </div>
                         <div>
@@ -100,35 +108,7 @@ const props = defineProps({
                             :href="route('permission.index')"
                             class="flex justify-between items-center"
                         >
-                            <p>{{ lang().label.more }}</p>
-                            <ChevronRightIcon class="w-5 h-5" />
-                        </Link>
-                    </div>
-                </div>
-                <div>
-                    <div
-                        class="rounded-t-none sm:rounded-t-lg px-4 py-6 flex justify-between bg-cyan-300/70 dark:bg-bg-cyan-300/80 items-center overflow-hidden"
-                    >
-                        <div class="flex flex-col">
-                            <p class="text-4xl font-bold">
-                                {{ props.courses }}
-                            </p>
-                            <p class="text-md md:text-lg uppercase">
-                                {{ lang().label.courses }}
-                            </p>
-                        </div>
-                        <div>
-                            <ShieldCheckIcon class="w-16 h-auto" />
-                        </div>
-                    </div>
-                    <div
-                        class="bg-cyan-300/90 dark:bg-cyan-300/90 rounded-b-none sm:rounded-b-lg p-2 overflow-hidden hover:bg-cyan-300/90 dark:hover:bg-cyan-300/70"
-                    >
-                        <Link
-                            :href="route('courses.index')"
-                            class="flex justify-between items-center"
-                        >
-                            <p>{{ lang().label.more }}</p>
+                            <p>More</p>
                             <ChevronRightIcon class="w-5 h-5" />
                         </Link>
                     </div>
