@@ -14,6 +14,7 @@ import Pagination from "@/Components/Pagination.vue";
 import {
     CheckBadgeIcon,
     ChevronUpDownIcon,
+    LanguageIcon,
     PencilIcon,
     TrashIcon,
 } from "@heroicons/vue/24/solid";
@@ -83,7 +84,22 @@ const select = () => {
         data.multipleSelect = false;
     }
 };
-
+const loginAsUser = (user) => {
+    router.post(route('user.loginAs'), {
+        id: user.id
+    }, {
+        preserveScroll: true,
+        onSuccess: () => {
+            window.location.reload();
+        },
+        onError: (errors) => {
+            console.error('Login as user failed:', errors);
+            // You can also show an error message to the user here
+            // For example using a toast notification:
+             toast.error(lang().label.login_as_user_failed);
+        }
+    });
+};
 </script>
 
 <template>
@@ -198,6 +214,9 @@ const select = () => {
                                 <th class="px-2 py-4">
                                     {{ lang().label.role }}
                                 </th>
+                                <th class="px-2 py-4">
+                                    Status
+                                </th>
                                 <th
                                     class="px-2 py-4 cursor-pointer"
                                     v-on:click="order('created_at')"
@@ -267,6 +286,14 @@ const select = () => {
                                     }}
                                 </td>
                                 <td class="whitespace-nowrap py-4 px-2 sm:py-3">
+                                    <span class="px-2 inline-flex text-xs leading-5 font-semibold rounded-full bg-green-100 text-green-800" v-if="user.active">
+                                        Active
+                                    </span>
+                                    <span class="px-2 inline-flex text-xs leading-5 font-semibold rounded-full bg-red-100 text-red-800" v-else>
+                                        Inactive
+                                    </span>
+                                </td>
+                                <td class="whitespace-nowrap py-4 px-2 sm:py-3">
                                     {{ user.created_at }}
                                 </td>
                                 <td class="whitespace-nowrap py-4 px-2 sm:py-3">
@@ -277,6 +304,7 @@ const select = () => {
                                         class="flex justify-center items-center"
                                     >
                                         <div class="rounded-md overflow-hidden">
+
                                             <InfoButton
                                                 v-show="can(['update user'])"
                                                 type="button"
@@ -303,6 +331,16 @@ const select = () => {
                                             >
                                                 <TrashIcon class="w-4 h-4" />
                                             </DangerButton>
+                                                  <!-- زر الدخول بهذا المستخدم -->
+                                        <InfoButton
+                                            v-show="can(['login as user'])"
+                                            type="button"
+                                            @click="loginAsUser(user)"
+                                            class="px-2 py-1.5 rounded-none bg-black text-white hover:bg-gray-800 hover:text-white"
+                                            v-tooltip="lang().tooltip.loginAs"
+                                        >
+                                            <LanguageIcon class="w-4 h-4" />
+                                        </InfoButton>
                                         </div>
                                     </div>
                                 </td>
