@@ -1,4 +1,4 @@
-const Ziggy = {"url":"http:\/\/localhost","port":null,"defaults":{},"routes":{"sanctum.csrf-cookie":{"uri":"sanctum\/csrf-cookie","methods":["GET","HEAD"]},"ignition.healthCheck":{"uri":"_ignition\/health-check","methods":["GET","HEAD"]},"ignition.executeSolution":{"uri":"_ignition\/execute-solution","methods":["POST"]},"ignition.updateConfig":{"uri":"_ignition\/update-config","methods":["POST"]},"setlang":{"uri":"setLang\/{locale}","methods":["GET","HEAD"],"parameters":["locale"]},"register":{"uri":"register","methods":["GET","HEAD"]},"login":{"uri":"login","methods":["GET","HEAD"]},"password.request":{"uri":"forgot-password","methods":["GET","HEAD"]},"password.email":{"uri":"forgot-password","methods":["POST"]},"password.reset":{"uri":"reset-password\/{token}","methods":["GET","HEAD"],"parameters":["token"]},"password.store":{"uri":"reset-password","methods":["POST"]},"verification.notice":{"uri":"verify-email","methods":["GET","HEAD"]},"verification.verify":{"uri":"verify-email\/{id}\/{hash}","methods":["GET","HEAD"],"parameters":["id","hash"]},"verification.send":{"uri":"email\/verification-notification","methods":["POST"]},"password.confirm":{"uri":"confirm-password","methods":["GET","HEAD"]},"password.update":{"uri":"password","methods":["PUT"]},"logout":{"uri":"logout","methods":["POST"]},"google.login":{"uri":"auth\/google","methods":["GET","HEAD"]},"dashboard":{"uri":"dashboard","methods":["GET","HEAD"]},"profile.edit":{"uri":"profile","methods":["GET","HEAD"]},"profile.update":{"uri":"profile","methods":["PATCH"]},"profile.destroy":{"uri":"profile","methods":["DELETE"]},"user.index":{"uri":"user","methods":["GET","HEAD"]},"user.store":{"uri":"user","methods":["POST"]},"user.update":{"uri":"user\/{user}","methods":["PUT","PATCH"],"parameters":["user"]},"user.destroy":{"uri":"user\/{user}","methods":["DELETE"],"parameters":["user"],"bindings":{"user":"id"}},"user.destroy-bulk":{"uri":"user\/destroy-bulk","methods":["POST"]},"role.index":{"uri":"role","methods":["GET","HEAD"]},"role.store":{"uri":"role","methods":["POST"]},"role.update":{"uri":"role\/{role}","methods":["PUT","PATCH"],"parameters":["role"],"bindings":{"role":"id"}},"role.destroy":{"uri":"role\/{role}","methods":["DELETE"],"parameters":["role"],"bindings":{"role":"id"}},"role.destroy-bulk":{"uri":"role\/destroy-bulk","methods":["POST"]},"permission.index":{"uri":"permission","methods":["GET","HEAD"]},"permission.store":{"uri":"permission","methods":["POST"]},"permission.update":{"uri":"permission\/{permission}","methods":["PUT","PATCH"],"parameters":["permission"],"bindings":{"permission":"id"}},"permission.destroy":{"uri":"permission\/{permission}","methods":["DELETE"],"parameters":["permission"],"bindings":{"permission":"id"}},"permission.destroy-bulk":{"uri":"permission\/destroy-bulk","methods":["POST"]},"courses.index":{"uri":"courses","methods":["GET","HEAD"]},"courses.store":{"uri":"courses","methods":["POST"]},"courses.update":{"uri":"courses\/{course}","methods":["PUT","PATCH"],"parameters":["course"]},"courses.destroy":{"uri":"{course}","methods":["DELETE"],"parameters":["course"]},"course.details":{"uri":"courses\/{id}\/details\/{courseSlug}","methods":["GET","HEAD"],"parameters":["id","courseSlug"]},"course.player":{"uri":"courses\/{courseId}\/player\/{courseSlug}","methods":["GET","HEAD"],"parameters":["courseId","courseSlug"]},"course.watchLecture":{"uri":"courses\/{courseId}\/player\/{courseSlug}\/watch\/{lectureID}","methods":["GET","HEAD"],"parameters":["courseId","courseSlug","lectureID"]},"lecture.markCompleted":{"uri":"lectures\/mark-completed","methods":["POST"]},"course.destroy-bulk":{"uri":"courses\/destroy-bulk","methods":["POST"]},"course.learn":{"uri":"courses\/{courseId}\/learn\/{courseSlug}","methods":["GET","HEAD"],"parameters":["courseId","courseSlug"]},"course.create.lecture":{"uri":"courses\/{courseId}\/lecture\/create","methods":["GET","HEAD"],"parameters":["courseId"]},"course.sections.store":{"uri":"courses\/{course}\/sections","methods":["POST"],"parameters":["course"],"bindings":{"course":"id"}},"course.sections.destroy":{"uri":"courses\/{course}\/sections\/{section}","methods":["DELETE"],"parameters":["course","section"]},"courses.edit":{"uri":"{course}\/edit","methods":["GET","HEAD"],"parameters":["course"]},"sections.edit":{"uri":"sections\/{id}\/edit","methods":["GET","HEAD"],"parameters":["id"]},"courses.enroll":{"uri":"courses\/{courseId}\/enroll","methods":["POST"],"parameters":["courseId"]},"courses.checkEnrollment":{"uri":"courses\/{courseId}\/check-enrollment","methods":["GET","HEAD"],"parameters":["courseId"]},"assessments.index":{"uri":"assessments","methods":["GET","HEAD"]},"assessments.create":{"uri":"assessments\/create","methods":["GET","HEAD"]},"assessments.reports":{"uri":"assessments\/reports","methods":["GET","HEAD"]},"assessments.store":{"uri":"assessments","methods":["POST"]},"assessments.show":{"uri":"assessments\/{assessment}","methods":["GET","HEAD"],"parameters":["assessment"]},"assessments.update":{"uri":"assessments\/{assessment}","methods":["PUT"],"parameters":["assessment"]},"assessments.destroy":{"uri":"assessments\/{assessment}","methods":["DELETE"],"parameters":["assessment"]},"admin.assessments.create":{"uri":"admin\/assessments\/create","methods":["GET","HEAD"]},"admin.assessments.store":{"uri":"admin\/assessments","methods":["POST"]},"admin.assessments.edit":{"uri":"admin\/assessments\/{assessment}\/edit","methods":["GET","HEAD"],"parameters":["assessment"]},"admin.assessments.update":{"uri":"admin\/assessments\/{assessment}","methods":["PUT"],"parameters":["assessment"]},"admin.assessments.destroy":{"uri":"admin\/assessments\/{assessment}","methods":["DELETE"],"parameters":["assessment"]},"admin.assessments.reports":{"uri":"admin\/assessments\/reports","methods":["GET","HEAD"]}}};
+const Ziggy = {"url":"http:\/\/localhost","port":null,"defaults":{},"routes":{"debugbar.openhandler":{"uri":"_debugbar\/open","methods":["GET","HEAD"]},"debugbar.clockwork":{"uri":"_debugbar\/clockwork\/{id}","methods":["GET","HEAD"],"parameters":["id"]},"debugbar.assets.css":{"uri":"_debugbar\/assets\/stylesheets","methods":["GET","HEAD"]},"debugbar.assets.js":{"uri":"_debugbar\/assets\/javascript","methods":["GET","HEAD"]},"debugbar.cache.delete":{"uri":"_debugbar\/cache\/{key}\/{tags?}","methods":["DELETE"],"parameters":["key","tags"]},"debugbar.queries.explain":{"uri":"_debugbar\/queries\/explain","methods":["POST"]},"sanctum.csrf-cookie":{"uri":"sanctum\/csrf-cookie","methods":["GET","HEAD"]},"log-viewer.hosts":{"uri":"log-viewer\/api\/hosts","methods":["GET","HEAD"]},"log-viewer.folders":{"uri":"log-viewer\/api\/folders","methods":["GET","HEAD"]},"log-viewer.folders.request-download":{"uri":"log-viewer\/api\/folders\/{folderIdentifier}\/download\/request","methods":["GET","HEAD"],"parameters":["folderIdentifier"]},"log-viewer.folders.clear-cache":{"uri":"log-viewer\/api\/folders\/{folderIdentifier}\/clear-cache","methods":["POST"],"parameters":["folderIdentifier"]},"log-viewer.folders.delete":{"uri":"log-viewer\/api\/folders\/{folderIdentifier}","methods":["DELETE"],"parameters":["folderIdentifier"]},"log-viewer.files":{"uri":"log-viewer\/api\/files","methods":["GET","HEAD"]},"log-viewer.files.request-download":{"uri":"log-viewer\/api\/files\/{fileIdentifier}\/download\/request","methods":["GET","HEAD"],"parameters":["fileIdentifier"]},"log-viewer.files.clear-cache":{"uri":"log-viewer\/api\/files\/{fileIdentifier}\/clear-cache","methods":["POST"],"parameters":["fileIdentifier"]},"log-viewer.files.delete":{"uri":"log-viewer\/api\/files\/{fileIdentifier}","methods":["DELETE"],"parameters":["fileIdentifier"]},"log-viewer.files.clear-cache-all":{"uri":"log-viewer\/api\/clear-cache-all","methods":["POST"]},"log-viewer.files.delete-multiple-files":{"uri":"log-viewer\/api\/delete-multiple-files","methods":["POST"]},"log-viewer.logs":{"uri":"log-viewer\/api\/logs","methods":["GET","HEAD"]},"log-viewer.folders.download":{"uri":"log-viewer\/api\/folders\/{folderIdentifier}\/download","methods":["GET","HEAD"],"parameters":["folderIdentifier"]},"log-viewer.files.download":{"uri":"log-viewer\/api\/files\/{fileIdentifier}\/download","methods":["GET","HEAD"],"parameters":["fileIdentifier"]},"log-viewer.index":{"uri":"log-viewer\/{view?}","methods":["GET","HEAD"],"wheres":{"view":"(.*)"},"parameters":["view"]},"ignition.healthCheck":{"uri":"_ignition\/health-check","methods":["GET","HEAD"]},"ignition.executeSolution":{"uri":"_ignition\/execute-solution","methods":["POST"]},"ignition.updateConfig":{"uri":"_ignition\/update-config","methods":["POST"]},"setlang":{"uri":"setLang\/{locale}","methods":["GET","HEAD"],"parameters":["locale"]},"register":{"uri":"register","methods":["GET","HEAD"]},"login":{"uri":"login","methods":["GET","HEAD"]},"password.request":{"uri":"forgot-password","methods":["GET","HEAD"]},"password.email":{"uri":"forgot-password","methods":["POST"]},"password.reset":{"uri":"reset-password\/{token}","methods":["GET","HEAD"],"parameters":["token"]},"password.store":{"uri":"reset-password","methods":["POST"]},"verification.notice":{"uri":"verify-email","methods":["GET","HEAD"]},"verification.verify":{"uri":"verify-email\/{id}\/{hash}","methods":["GET","HEAD"],"parameters":["id","hash"]},"verification.send":{"uri":"email\/verification-notification","methods":["POST"]},"password.confirm":{"uri":"confirm-password","methods":["GET","HEAD"]},"password.update":{"uri":"password","methods":["PUT"]},"logout":{"uri":"logout","methods":["POST"]},"google.login":{"uri":"auth\/google","methods":["GET","HEAD"]},"user.loginAs":{"uri":"user\/login-as","methods":["POST"]},"courses.index":{"uri":"courses","methods":["GET","HEAD"]},"courses.details":{"uri":"courses\/{id}\/details\/{courseSlug}","methods":["GET","HEAD"],"parameters":["id","courseSlug"]},"courses.learn":{"uri":"courses\/{courseId}\/learn\/{courseSlug}","methods":["GET","HEAD"],"parameters":["courseId","courseSlug"]},"courses.enroll":{"uri":"courses\/{courseId}\/enroll","methods":["POST"],"parameters":["courseId"]},"courses.player":{"uri":"courses\/{courseId}\/player\/{courseSlug}","methods":["GET","HEAD"],"parameters":["courseId","courseSlug"]},"courses.watch":{"uri":"courses\/{courseId}\/player\/{courseSlug}\/watch\/{lectureID}","methods":["GET","HEAD"],"parameters":["courseId","courseSlug","lectureID"]},"courses.lecture.complete":{"uri":"courses\/lectures\/mark-completed","methods":["POST"]},"dashboard":{"uri":"dashboard","methods":["GET","HEAD"]},"profile.edit":{"uri":"profile","methods":["GET","HEAD"]},"profile.update":{"uri":"profile","methods":["PATCH"]},"profile.destroy":{"uri":"profile","methods":["DELETE"]},"user.index":{"uri":"user","methods":["GET","HEAD"]},"user.store":{"uri":"user","methods":["POST"]},"user.update":{"uri":"user\/{}","methods":["PUT","PATCH"],"parameters":[""]},"user.destroy":{"uri":"user\/{}","methods":["DELETE"],"parameters":[""]},"user.destroy-bulk":{"uri":"user\/destroy-bulk","methods":["POST"]},"role.index":{"uri":"role","methods":["GET","HEAD"]},"role.store":{"uri":"role","methods":["POST"]},"role.update":{"uri":"role\/{role}","methods":["PUT","PATCH"],"parameters":["role"],"bindings":{"role":"id"}},"role.destroy":{"uri":"role\/{role}","methods":["DELETE"],"parameters":["role"],"bindings":{"role":"id"}},"destroy-bulk":{"uri":"destroy-bulk","methods":["POST"]},"permission.index":{"uri":"permission","methods":["GET","HEAD"]},"permission.store":{"uri":"permission","methods":["POST"]},"permission.update":{"uri":"permission\/{}","methods":["PUT","PATCH"],"parameters":[""]},"permission.destroy":{"uri":"permission\/{}","methods":["DELETE"],"parameters":[""]},"permission.destroy-bulk":{"uri":"permission\/destroy-bulk","methods":["POST"]},"courses.create":{"uri":"courses\/create","methods":["GET","HEAD"]},"courses.store":{"uri":"courses","methods":["POST"]},"courses.destroy-bulk":{"uri":"courses\/destroy-bulk","methods":["POST"]},"courses.details.show":{"uri":"courses\/{courseId}\/details","methods":["GET","HEAD"],"parameters":["courseId"]},"courses.courses.edit":{"uri":"courses\/{course}\/edit","methods":["GET","HEAD"],"parameters":["course"]},"courses.lecture.create":{"uri":"courses\/{course}\/lecture\/create","methods":["GET","HEAD"],"parameters":["course"],"bindings":{"course":"id"}},"courses.sections.store":{"uri":"courses\/{course}\/sections","methods":["POST"],"parameters":["course"],"bindings":{"course":"id"}},"courses.sections.lectures.store":{"uri":"courses\/{course}\/sections\/{section}\/lectures","methods":["POST"],"parameters":["course","section"],"bindings":{"course":"id"}},"courses.sections.destroy":{"uri":"courses\/{course}\/sections\/{section}","methods":["DELETE"],"parameters":["course","section"]},"courses.quizzes":{"uri":"courses\/{course}\/quizzes","methods":["GET","HEAD"],"parameters":["course"],"bindings":{"course":"id"}},"courses.check-enrollment":{"uri":"courses\/{courseId}\/check-enrollment","methods":["GET","HEAD"],"parameters":["courseId"]},"courses.update-progress":{"uri":"courses\/{courseId}\/update-progress","methods":["POST"],"parameters":["courseId"]},"quizzes.index":{"uri":"quizzes","methods":["GET","HEAD"]},"quizzes.create":{"uri":"quizzes\/create","methods":["GET","HEAD"]},"quizzes.store":{"uri":"quizzes","methods":["POST"]},"quizzes.reports":{"uri":"quizzes\/reports","methods":["GET","HEAD"]},"quizzes.show":{"uri":"quizzes\/{quiz}","methods":["GET","HEAD"],"parameters":["quiz"],"bindings":{"quiz":"id"}},"quizzes.update":{"uri":"quizzes\/{quiz}","methods":["PUT"],"parameters":["quiz"],"bindings":{"quiz":"id"}},"quizzes.destroy":{"uri":"quizzes\/{quiz}","methods":["DELETE"],"parameters":["quiz"],"bindings":{"quiz":"id"}},"quizzes.questions.list":{"uri":"quizzes\/{quiz}\/questions","methods":["GET","HEAD"],"parameters":["quiz"],"bindings":{"quiz":"id"}},"quizzes.questions.create":{"uri":"quizzes\/{quiz}\/questions\/create","methods":["GET","HEAD"],"parameters":["quiz"],"bindings":{"quiz":"id"}},"quizzes.questions.store":{"uri":"quizzes\/{quiz}\/questions","methods":["POST"],"parameters":["quiz"],"bindings":{"quiz":"id"}},"quizzes.start":{"uri":"quizzes\/{quiz}\/start","methods":["POST"],"parameters":["quiz"],"bindings":{"quiz":"id"}},"quizzes.take":{"uri":"quizzes\/{quiz}\/take\/{attempt}","methods":["GET","HEAD"],"parameters":["quiz","attempt"],"bindings":{"quiz":"id","attempt":"id"}},"quizzes.submit":{"uri":"quizzes\/{quiz}\/submit\/{attempt}","methods":["POST"],"parameters":["quiz","attempt"],"bindings":{"quiz":"id","attempt":"id"}},"quizzes.import-questions":{"uri":"quizzes\/{quiz}\/import-questions","methods":["POST"],"parameters":["quiz"],"bindings":{"quiz":"id"}},"quizzes.import-questions-json":{"uri":"quizzes\/{quiz}\/import-questions-json","methods":["POST"],"parameters":["quiz"],"bindings":{"quiz":"id"}},"quizzes.template.download":{"uri":"quizzes\/template\/download","methods":["GET","HEAD"]},"admin.quizzes.create":{"uri":"admin\/quizzes\/create","methods":["GET","HEAD"]},"admin.quizzes.store":{"uri":"admin\/quizzes","methods":["POST"]},"admin.quizzes.edit":{"uri":"admin\/quizzes\/{quiz}\/edit","methods":["GET","HEAD"],"parameters":["quiz"],"bindings":{"quiz":"id"}},"admin.quizzes.update":{"uri":"admin\/quizzes\/{quiz}","methods":["PUT"],"parameters":["quiz"],"bindings":{"quiz":"id"}},"admin.quizzes.destroy":{"uri":"admin\/quizzes\/{quiz}","methods":["DELETE"],"parameters":["quiz"],"bindings":{"quiz":"id"}},"admin.quizzes.reports":{"uri":"admin\/quizzes\/reports","methods":["GET","HEAD"]}}};
 
 if (typeof window !== 'undefined' && typeof window.Ziggy !== 'undefined') {
     Object.assign(Ziggy.routes, window.Ziggy.routes);
