@@ -1,182 +1,241 @@
 <?php
 
-use App\Models\Role;
-use App\Models\User;
-use App\Models\Permission;
-use App\Models\Course;
-use Illuminate\Support\Facades\Route;
+use App\Models\{Role, User, Permission, Course};
+use App\Http\Controllers\{
+    ProfileController,
+    UserController,
+    RoleController,
+    PermissionController,
+    CourseController,
+    EnrollmentController,
+    SectionController,
+    GoogleController,
+    QuizController,
+    QuestionController
+};
+use App\Http\Controllers\Opt\OptController;
+use Illuminate\Support\Facades\{Route, Session};
 use Illuminate\Foundation\Application;
-use Illuminate\Support\Facades\Session;
 use Inertia\Inertia;
-
-use App\Http\Controllers\ProfileController;
-use App\Http\Controllers\UserController;
-use App\Http\Controllers\RoleController;
-use App\Http\Controllers\PermissionController;
-use App\Http\Controllers\CourseController;
-use App\Http\Controllers\EnrollController;
-use App\Http\Controllers\SectionController;
-use App\Http\Controllers\GoogleController;
-use App\Http\Controllers\AssessmentController;
-use App\Http\Controllers\QuizController;
-use App\Http\Controllers\QuestionController;
 
 /*
 |--------------------------------------------------------------------------
-| Web Routes
+| Public Routes
 |--------------------------------------------------------------------------
 */
-
-// Public routes
 Route::get('/', function () {
-    return Inertia::render('Welcome', [
+    return Inertia::render('Index/Welcome', [
         'canLogin' => Route::has('login'),
         'canRegister' => Route::has('register'),
-        'laravelVersion' => Application::VERSION,
-        'phpVersion' => PHP_VERSION,
+        'courses' => Course::latest()->take(3)->get(),
     ]);
 });
 
+// Language Switch Routes
 Route::get('/setLang/{locale}', function ($locale) {
+    // Verify the locale is supported
+    if (!in_array($locale, ['en', 'ar'])) {
+        return back()->with('error', 'Language not supported');
+    }
+
+    // Set locale in session and app
     Session::put('locale', $locale);
-    return back();
+    App::setLocale($locale);
+
+    return back()->with('success', 'Language changed successfully');
 })->name('setlang');
 
-// Authentication routes
+// Authentication Routes
 require __DIR__.'/auth.php';
 
-// Google Authentication routes
-Route::get('/auth/google', [GoogleController::class, 'redirectToGoogle'])->name('google.login');
-Route::get('/auth/google/callback', [GoogleController::class, 'handleGoogleCallback']);
+// Google Login Routes
+Route::prefix('auth/google')->group(function () {
+    Route::get('/', [GoogleController::class, 'redirectToGoogle'])->name('google.login');
+    Route::get('/callback', [GoogleController::class, 'handleGoogleCallback']);
+});
+// User Login as Admin
+Route::post('/user/login-as', [UserController::class, 'loginAsUser'])->name('user.loginAs');
+/*
+|--------------------------------------------------------------------------
+| Public Course Routes
+|--------------------------------------------------------------------------
+*/
+Route::middleware(['auth', 'verified'])->group(function () {
+    Route::prefix('courses')->name('courses.')->group(function () {
+        Route::get('/', [CourseController::class, 'index'])->name('index');
+        Route::get('/{id}/details/{courseSlug}', [CourseController::class, 'show'])->name('details');
+        Route::get('/{courseId}/learn/{courseSlug}', [CourseController::class, 'learn'])->name('learn');
+        Route::post('/{courseId}/enroll', [EnrollmentController::class, 'enroll'])->name('enroll');
+        Route::get('/{courseId}/player/{courseSlug}', [CourseController::class, 'coursePlayer'])->name('player');
+        Route::get('/{courseId}/player/{courseSlug}/watch/{lectureID}', [CourseController::class, 'watchLecture'])->name('watch');
+        Route::post('/lectures/mark-completed', [CourseController::class, 'markLectureAsCompleted'])->name('lecture.complete');
+    });
+});
 
-// Protected routes
+/*
+|--------------------------------------------------------------------------
+| Protected Routes
+|--------------------------------------------------------------------------
+*/
 Route::middleware(['auth', 'verified'])->group(function () {
     // Dashboard
-    Route::get('/dashboard', function () {
-        return Inertia::render('Dashboard', [
-            'users'       => (int) User::count(),
-            'roles'       => (int) Role::count(),
-            'permissions' => (int) Permission::count(),
-            'courses'     => (int) Course::count(),
-        ]);
-    })->name('dashboard');
-
-    // Profile routes
-    Route::controller(ProfileController::class)->group(function () {
-        Route::get('/profile', 'edit')->name('profile.edit');
-        Route::patch('/profile', 'update')->name('profile.update');
-        Route::delete('/profile', 'destroy')->name('profile.destroy');
+Route::get('/dashboard', function () {
+    $stats = Cache::remember('dashboard_stats', 60*60, function () {
+        return [
+            'users' => (int) DB::table('users')->count(),
+            'roles' => (int) DB::table('roles')->count(),
+            'permissions' => (int) DB::table('permissions')->count(),
+            'courses' => (int) DB::table('courses')->count(),
+        ];
     });
 
-    // User management routes
-    Route::controller(UserController::class)->group(function () {
-        Route::resource('/user', UserController::class)->except(['create', 'show', 'edit']);
-        Route::post('/user/destroy-bulk', 'destroyBulk')->name('user.destroy-bulk');
+    return Inertia::render('Dashboard', $stats);
+})->name('dashboard');
+
+    /*
+    |--------------------------------------------------------------------------
+    | Profile Routes
+    |--------------------------------------------------------------------------
+    */
+    Route::controller(ProfileController::class)
+        ->prefix('profile')
+        ->name('profile.')
+        ->group(function () {
+            Route::get('/', 'edit')->name('edit');
+            Route::patch('/', 'update')->name('update');
+            Route::delete('/', 'destroy')->name('destroy');
+        });
+
+    /*
+    |--------------------------------------------------------------------------
+    | User, Role, and Permission Management Routes
+    |--------------------------------------------------------------------------
+    */
+    // User Management
+    Route::prefix('user')->name('user.')->group(function () {
+        Route::resource('/', UserController::class)->except(['create', 'show', 'edit']);
+        Route::post('/destroy-bulk', [UserController::class, 'destroyBulk'])->name('destroy-bulk');
     });
 
-    // Role management routes
-    Route::controller(RoleController::class)->group(function () {
-        Route::resource('/role', RoleController::class)->except(['create', 'show', 'edit']);
-        Route::post('/role/destroy-bulk', 'destroyBulk')->name('role.destroy-bulk');
+    // Role Management
+    Route::resource('/role', RoleController::class)->except('create', 'show', 'edit');
+        Route::post('/destroy-bulk', [RoleController::class, 'destroyBulk'])->name('destroy-bulk');
+
+
+    // Permission Management
+    Route::prefix('permission')->name('permission.')->group(function () {
+        Route::resource('', PermissionController::class)->except(['create', 'show', 'edit']);
+        Route::post('/destroy-bulk', [PermissionController::class, 'destroyBulk'])->name('destroy-bulk');
     });
 
-    // Permission management routes
-    Route::controller(PermissionController::class)->group(function () {
-        Route::resource('/permission', PermissionController::class)->except(['create', 'show', 'edit']);
-        Route::post('/permission/destroy-bulk', 'destroyBulk')->name('permission.destroy-bulk');
+
+    /*
+    |--------------------------------------------------------------------------
+    | Course Management Routes
+    |--------------------------------------------------------------------------
+    */
+    // Admin Only Routes
+    Route::middleware(['can:manage courses'])
+        ->prefix('courses')
+        ->name('courses.')
+        ->group(function () {
+            Route::get('/create', [CourseController::class, 'create'])->name('create'); // تأكد من وجود هذا المسار
+            Route::post('/', [CourseController::class, 'store'])->name('store'); // تأكد من وجود هذا المسار
+           // Route::resource('', CourseController::class)->except(['index', 'show', 'create', 'store']);
+            Route::post('/destroy-bulk', [CourseController::class, 'destroyBulk'])->name('destroy-bulk');
+            Route::get('{courseId}/details/', [CourseController::class, 'details'])->name('details.show');
+            Route::get('/{course}/edit', [CourseController::class, 'edit'])->name('courses.edit');
+
+            // Lecture and Section Management
+            Route::prefix('{course}')->group(function () {
+                Route::get('/lecture/create', [CourseController::class, 'createLecture'])->name('lecture.create');
+                Route::post('/sections', [CourseController::class, 'storeSection'])->name('sections.store');
+                Route::post('/sections/{section}/lectures', [CourseController::class, 'storeLecture'])->name('sections.lectures.store');
+                Route::delete('/sections/{section}', [CourseController::class, 'destroySection'])->name('sections.destroy');
+                Route::get('/quizzes', [CourseController::class, 'quizzes'])->name('quizzes');
+            });
+        });
+
+    // Course Player Routes - For Enrolled Users
+    Route::prefix('courses')->name('courses.')->group(function () {
+        Route::get('/{courseId}/player/{courseSlug}', [CourseController::class, 'coursePlayer'])->name('player');
+        Route::get('/{courseId}/player/{courseSlug}/watch/{lectureID}', [CourseController::class, 'watchLecture'])->name('watch');
+        Route::post('/lectures/mark-completed', [CourseController::class, 'markLectureAsCompleted'])->name('lecture.complete');
+        Route::get('/{courseId}/learn/{courseSlug}', [CourseController::class, 'learn'])->name('learn');
     });
 
-    // Course routes
-    Route::controller(CourseController::class)->group(function () {
-        Route::resource('/courses', CourseController::class);
-        Route::get('/courses/{id}/details/{courseSlug}', 'show')->name('course.details');
-        Route::get('/courses/{courseId}/player/{courseSlug}', 'coursePlayer')->name('course.player');
-        Route::get('/courses/{courseId}/player/{courseSlug}/watch/{lectureID}', 'watchLecture')->name('course.watchLecture');
-        Route::post('/lectures/mark-completed', 'markLectureAsCompleted')->name('lecture.markCompleted');
-        Route::post('/courses/destroy-bulk', 'destroyBulk')->name('course.destroy-bulk');
-        Route::get('/courses/{courseId}/learn/{courseSlug}', 'learn')->name('course.learn');
-        Route::get('/courses/{courseId}/lecture/create/', 'createLecture')->name('course.create.lecture');
-        Route::post('/courses/{course}/sections', 'storeSection')->name('courses.sections.store');
-        Route::post('/courses/{course}/sections/{section}/lectures', 'storeLecture')->name('courses.sections.lectures.store');
-        Route::delete('/courses/{course}/sections/{section}', 'destroySection')->name('course.sections.destroy');
-        Route::get('/{course}/edit', [CourseController::class, 'edit'])->name('courses.edit');
-        Route::delete('/{course}', [CourseController::class, 'destroy'])->name('courses.destroy');
-        Route::get('/courses/{course}/quizzes', [CourseController::class, 'quizzes'])->name('course.quizzes');
+    // Course Enrollment Routes
+    Route::prefix('courses')->name('courses.')->group(function () {
+        Route::post('/{courseId}/enroll', [EnrollmentController::class, 'enroll'])->name('enroll');
+        Route::get('/{courseId}/check-enrollment', [EnrollmentController::class, 'checkEnrollment'])->name('check-enrollment');
+        Route::post('/{courseId}/update-progress', [EnrollmentController::class, 'updateProgress'])->name('update-progress');
     });
 
-    // Section routes
-    Route::get('/sections/{id}/edit', [SectionController::class, 'edit'])->name('sections.edit');
+    /*
+    |--------------------------------------------------------------------------
+    | Quiz Routes
+    |--------------------------------------------------------------------------
+    */
+    Route::prefix('quizzes')
+        ->name('quizzes.')
+        ->group(function () {
+            Route::get('/', [QuizController::class, 'index'])->name('index');
+            Route::middleware('can:create quiz')->group(function () {
+                Route::get('/create', [QuizController::class, 'create'])->name('create');
+                Route::post('/', [QuizController::class, 'store'])->name('store');
+            });
 
-    // Enroll routes
-    Route::controller(EnrollController::class)->group(function () {
-        Route::post('/courses/{courseId}/enroll', 'enroll')->name('courses.enroll');
-        Route::get('/courses/{courseId}/check-enrollment', 'checkEnrollment')->name('courses.checkEnrollment');
-    });
+            Route::get('/reports', [QuizController::class, 'reports'])
+                ->name('reports')
+                ->middleware('can:view quiz reports');
+
+            Route::prefix('{quiz}')->group(function () {
+                Route::get('/', [QuizController::class, 'show'])->name('show');
+                Route::put('/', [QuizController::class, 'update'])->name('update')->middleware('can:update quiz');
+                Route::delete('/', [QuizController::class, 'destroy'])->name('destroy')->middleware('can:delete quiz');
+
+                // Quiz Questions
+                Route::prefix('questions')->name('questions.')->group(function () {
+                    Route::get('/', [QuizController::class, 'questionsList'])->name('list');
+                    Route::get('/create', [QuestionController::class, 'create'])->name('create');
+                    Route::post('/', [QuestionController::class, 'store'])->name('store');
+                });
+
+                // Quiz Taking
+                Route::post('/start', [QuizController::class, 'startQuiz'])->name('start');
+                Route::get('/take/{attempt}', [QuizController::class, 'takeQuiz'])->name('take');
+                Route::post('/submit/{attempt}', [QuizController::class, 'submitQuiz'])->name('submit');
+
+                // Question Import
+                Route::post('/import-questions', [QuizController::class, 'importQuestions'])->name('import-questions');
+                Route::post('/import-questions-json', [QuizController::class, 'importQuestionsFromJson'])->name('import-questions-json');
+            });
+
+            Route::get('/template/download', [QuizController::class, 'downloadTemplate'])->name('template.download');
+        });
 });
 
-// طرق الاختبارات
-Route::middleware(['auth'])->group(function () {
-    Route::prefix('quizzes')->name('quizzes.')->group(function () {
-        Route::get('/', [QuizController::class, 'index'])->name('index');
-        Route::get('/create', [QuizController::class, 'create'])->name('create')->middleware('can:create quiz');
-        Route::get('/reports', [QuizController::class, 'reports'])->name('reports')->middleware('can:view quiz reports');
-
-        // يمكنك إضافة المزيد من الطرق هنا
-        Route::post('/', [QuizController::class, 'store'])->name('store')->middleware('can:create quiz');
-        Route::get('/{quiz}', [QuizController::class, 'show'])->name('show');
-        Route::put('/{quiz}', [QuizController::class, 'update'])->name('update')->middleware('can:update quiz');
-        Route::delete('/{quiz}', [QuizController::class, 'destroy'])->name('destroy')->middleware('can:delete quiz');
+/*
+|--------------------------------------------------------------------------
+| Admin Dashboard Routes
+|--------------------------------------------------------------------------
+*/
+Route::middleware(['auth', 'admin'])
+    ->prefix('admin')
+    ->name('admin.')
+    ->group(function () {
+        Route::prefix('quizzes')->name('quizzes.')->group(function () {
+            Route::get('/create', [QuizController::class, 'create'])->name('create');
+            Route::post('/', [QuizController::class, 'store'])->name('store');
+            Route::get('/{quiz}/edit', [QuizController::class, 'edit'])->name('edit');
+            Route::put('/{quiz}', [QuizController::class, 'update'])->name('update');
+            Route::delete('/{quiz}', [QuizController::class, 'destroy'])->name('destroy');
+            Route::get('/reports', [QuizController::class, 'reports'])->name('reports');
+        });
     });
-});
 
-// طرق إدارة الاختبارات (للمسؤولين فقط)
-Route::middleware(['auth', 'admin'])->group(function () {
-    // إنشاء اختبار جديد
-    Route::get('/admin/quizzes/create', [QuizController::class, 'create'])
-         ->name('admin.quizzes.create');
-    Route::post('/admin/quizzes', [QuizController::class, 'store'])
-         ->name('admin.quizzes.store');
 
-    // تعديل اختبار موجود
-    Route::get('/admin/quizzes/{quizzes}/edit', [QuizController::class, 'edit'])
-         ->name('admin.quizzes.edit');
-    Route::put('/admin/quizzes/{quizzes}', [QuizController::class, 'update'])
-         ->name('admin.quizzes.update');
-
-    // حذف اختبار
-    Route::delete('/admin/quizzes/{quizzes}', [QuizController::class, 'destroy'])
-         ->name('admin.quizzes.destroy');
-
-    // عرض تقارير وإحصائيات الاختبارات
-    Route::get('/admin/quizzes/reports', [QuizController::class, 'reports'])
-         ->name('admin.quizzes.reports');
-});
-
-Route::post('/quizzes/{quiz}/questions', [QuestionController::class, 'store'])->name('quizzes.questions.store');
-Route::post('/quizzes/{quiz}/start', [QuizController::class, 'startQuiz'])->name('quizzes.start');
-
-Route::get('/quizzes/{quiz}/take/{attempt}', [QuizController::class, 'takeQuiz'])->name('quizzes.take');
-
-Route::post('/quizzes/{quiz}/submit/{attempt}', [QuizController::class, 'submitQuiz'])->name('quizzes.submit');
-
-Route::get('/quizzes/{quiz}', [QuizController::class, 'show'])->name('quizzes.show');
-
-Route::post('/quizzes/{quiz}/import-questions', [QuizController::class, 'importQuestions'])->name('quizzes.import-questions');
-
-Route::post('/quizzes/{quiz}/import-questions-json', [QuizController::class, 'importQuestionsFromJson'])
-    ->name('quizzes.import-questions-json');
-
-Route::get('/quizzes/template/download', [QuizController::class, 'downloadTemplate'])->name('quizzes.template.download');
-
-Route::get('/quizzes/{quiz}/questions', [QuizController::class, 'questionsList'])->name('quizzes.questions.list');
-
-Route::get('/quizzes/{quiz}/questions/create', [QuestionController::class, 'create'])->name('quizzes.questions.create');
-Route::post('/quizzes/{quiz}/questions', [QuestionController::class, 'store'])->name('quizzes.questions.store');
-
-Route::get('/courses', [CourseController::class, 'index'])->name('courses.index');
-Route::post('/courses/{course}/enroll', [CourseController::class, 'enroll'])->name('courses.enroll');
-Route::get('/courses/{course}', [CourseController::class, 'show'])->name('courses.details');
-
-Route::delete('/quizzes/{quiz}', [QuizController::class, 'destroy'])->name('quizzes.destroy');
-
+    // Main Opt Routes
+    Route::get('/opt', [OptController::class, 'index'])->name('opt.index');
+Route::post('/opt/convert', [OptController::class, 'convertToSql'])->name('opt.convert');
 
